requirements.txt
# Unified dependencies from interview + interview-tts
# Merge strategy: keep interview version when conflict, add missing packages from interview-tts
# Note: TTS-specific packages have been moved to requirements-tts.txt

absl-py==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
async-timeout==5.0.1
attrs==26.1.0
babel==2.18.0
bcrypt==5.0.0
bidict==0.23.1
blinker==1.9.0
boto3==1.42.78
botocore==1.42.78
brotli==1.2.0
build==1.4.1
cached_path==1.8.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
chromadb==1.5.5
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cryptography==46.0.5
cycler==0.12.1
dashscope==1.25.15
qwen-asr==0.0.6
dateparser==1.1.8
decorator==5.2.1
Deprecated==1.3.1
dnspython==2.8.0
docopt==0.6.2
durationpy==0.10
eventlet==0.33.3
exceptiongroup==1.3.1
fastapi==0.135.2
filelock==3.25.2
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.4
flatbuffers==25.12.19
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
google-api-core==2.30.0
googleapis-common-protos==1.73.0
google-auth==2.49.1
google-cloud-core==2.5.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-resumable-media==2.8.0
gradio_client==2.4.0
gradio==6.10.0
greenlet==3.3.2
groovy==0.1.2
grpcio==1.78.0
h11==0.16.0
hf-gradio==0.3.0
hf-xet==1.4.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==0.36.2
humanfriendly==10.0
idna==3.11
importlib_metadata==8.4.0
importlib_resources==6.5.2
inflect==7.0.0
itsdangerous==2.2.0
jax==0.6.2
jaxlib==0.6.2
jieba==0.42.1
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
jsonlines==1.2.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
kubernetes==35.0.0
langid==1.1.6
llvmlite==0.46.0
loguru==0.7.2
lxml==6.0.2
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.14
ml_dtypes==0.5.4
mmh3==5.2.1
mpmath==1.3.0
multidict==6.7.1
networkx==2.8.8
nltk==3.9.4
numba==0.64.0
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.23.2
opencv-contrib-python==4.8.1.78
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opt_einsum==3.4.0
orjson==3.11.7
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pdf2image==1.17.0
Pillow==10.0.1
platformdirs==4.9.4
pooch==1.9.0
propcache==0.4.1
protobuf==4.25.9
proto-plus==1.27.2
psutil==5.9.8
pyasn1_modules==0.4.2
pyasn1==0.6.3
pybase64==1.4.3
pycparser==3.0
pydantic_core==2.41.5
pydantic==2.12.5
pydantic-settings==2.13.1
Pygments==2.19.2
pyparsing==3.3.2
pypdf==6.9.2
PyPika==0.51.1
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-engineio==4.7.1
python-multipart==0.0.22
python-socketio==5.9.0
pytz==2026.1.post1
PyYAML==6.0.1
referencing==0.37.0
regex==2026.2.28
reportlab==4.0.5
requests==2.32.5
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.30.0
s3transfer==0.16.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
semantic-version==2.10.0
sentence-transformers==5.3.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
starlette==0.52.1
sympy==1.14.0
tabulate==0.10.0
tenacity==9.1.4
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.22.2
tomli==2.4.0
tomlkit==0.13.3
torch==2.11.0
torchaudio==2.11.0
tqdm==4.67.3
transformers==4.57.6
typer==0.24.1
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.42.0
wasabi==0.10.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.7
win32_setctime==1.2.0
wrapt==2.1.2
wsproto==1.3.2
yarl==1.23.0
zipp==3.23.0
